chore: Update azsecrets to v0.13.0 and the new/moved module
